% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DuckDBSpatial-package.R
\docType{package}
\name{DuckDBSpatial-package}
\alias{DuckDBSpatial}
\alias{DuckDBSpatial-package}
\title{DuckDBSpatial: DuckDB-Backed Spatial Data Structures}
\description{
Provides DuckDB-backed implementations of spatial data structures for efficient out-of-memory operations on large spatial datasets. The package includes 68 sf-compatible spatial methods (st_intersects, st_contains, st_within, etc.) and GeoParquet 1.0 I/O support via writeGeoParquet(). Ideal for spatial analysis, genomic interval analysis, and integration with BiocDuckDB for MultiAssaySpatialExperiment workflows requiring memory-efficient operations on large datasets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Genentech/DuckDBSpatial}
  \item Report bugs at \url{https://github.com/Genentech/DuckDBSpatial/issues}
}

}
\author{
\strong{Maintainer}: Patrick Aboyoun \email{aboyounp@gene.com}

Authors:
\itemize{
  \item Patrick Aboyoun \email{aboyounp@gene.com}
}

Other contributors:
\itemize{
  \item Genentech, Inc. [copyright holder]
}

}
\keyword{internal}
