% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DuckDBDataFrame-spatial.R
\name{layerSubsetByGeometry}
\alias{layerSubsetByGeometry}
\title{Row indices of a layer intersecting a geometry}
\usage{
layerSubsetByGeometry(x, y, coords = NULL, geom = "geometry")
}
\arguments{
\item{x}{A \code{DuckDBDataFrame} or in-memory tabular layer.}

\item{y}{A geometry to test against.}

\item{coords}{Optional x/y column names for point layers.}

\item{geom}{Geometry column name for polygon layers.}
}
\value{
Integer row indices.
}
\description{
Row indices of a layer intersecting a geometry
}
\examples{
pts_path <- tempfile(fileext = ".csv")
on.exit(unlink(pts_path), add = TRUE)
write.csv(data.frame(x = c(1, 5, 10), y = c(1, 5, 10)), pts_path, row.names = FALSE)
pts <- DuckDBDataFrame(pts_path, datacols = c("x", "y"))
poly <- sf::st_as_sfc("POLYGON((0 0, 6 0, 6 6, 0 6, 0 0))")
layerSubsetByGeometry(pts, poly, coords = c("x", "y"))

}
