% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoparquet-io.R
\name{readGeoParquet}
\alias{readGeoParquet}
\title{Read a GeoParquet file as a lazy \code{DuckDBDataFrame}}
\usage{
readGeoParquet(path, ...)
}
\arguments{
\item{path}{Path to a GeoParquet file or directory.}

\item{...}{Passed to \code{\link[DuckDBDataFrame]{DuckDBDataFrame}}.}
}
\value{
A \link[DuckDBDataFrame:DuckDBDataFrame-class]{DuckDBDataFrame}.
}
\description{
Read a GeoParquet file as a lazy \code{DuckDBDataFrame}
}
